kvserver/rangefeed: move helpers to processor_helpers_test
This patch moves the helper functions in processor_test.go to
processor_helpers_test.go.

Epic: none
Release note: none
